% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{get_kernel}
\alias{get_kernel}
\title{Obtain 2D kernel density estimates given sufficient statistics for lambdas
and the initial data Y0}
\usage{
get_kernel(lambdas, sigma2, dens_grid, N, T, Y0)
}
\arguments{
\item{lambdas}{sufficient statistics for the intercept term}

\item{sigma2}{variance of the shocks}

\item{dens_grid}{grid over which the density is to be computed}

\item{N}{cross-sectional dimension of the data}

\item{T}{time dimension of the data}

\item{Y0}{initial observations of the dependent variable}
}
\description{
Obtain 2D kernel density estimates given sufficient statistics for lambdas
and the initial data Y0
}
