% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{get_lambda0}
\alias{get_lambda0}
\title{Produce sufficient statistics (lambda0) given the common coefficients (rho0)}
\usage{
get_lambda0(rho, alpha = rep(0, n_alpha), N, T, n_alpha, Y_mat, X_mat, W,
  Z_mat)
}
\arguments{
\item{rho}{lagged dependent variable coefficients}

\item{alpha}{external variables coefficients}

\item{N}{cross-sectional dimension of the data}

\item{T}{time dimension of the data}

\item{n_alpha}{number of external variables}

\item{Y_mat}{dependent variable matrix}

\item{X_mat}{lagged dependent variable matrix}

\item{W}{cross-sectionally invariant variables - not used now}

\item{Z_mat}{external variable matrix}
}
\description{
Produce sufficient statistics (lambda0) given the common coefficients (rho0)
}
