% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{get_sigma2}
\alias{get_sigma2}
\title{Produce variance of the shocks estimated using GMM residues (sigma2_0)
given the common coefficients (rho0)}
\usage{
get_sigma2(rho, alpha = 0, common_par_method, X_star, Y_star, Z_star,
  X_mat, Y_mat, Z_mat, n_alpha)
}
\arguments{
\item{rho}{lagged dependent variable coefficients}

\item{alpha}{external variables coefficients}

\item{common_par_method}{method for estimating common parameters}

\item{X_star}{auxiliary matrix for OFD transformation}

\item{Y_star}{auxiliary matrix for OFD transformation}

\item{Z_star}{auxiliary matrix for OFD transformation}

\item{X_mat}{lagged dependent variable matrix}

\item{Y_mat}{dependent variable matrix}

\item{Z_mat}{external variable matrix}

\item{n_alpha}{number of external variables}
}
\description{
Produce variance of the shocks estimated using GMM residues (sigma2_0)
given the common coefficients (rho0)
}
