% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{loglikelihood_QMLE}
\alias{loglikelihood_QMLE}
\title{Produce (negative) log marginal likelihood for QMLE with correlated random
coefficients}
\usage{
loglikelihood_QMLE(param, n_alpha, X_mat, Y_mat, Z_mat, W, T, N, aux_Y0)
}
\arguments{
\item{param}{vectores of parameters to optimize over}

\item{n_alpha}{number of external variables}

\item{X_mat}{lagged dependent variable matrix}

\item{Y_mat}{dependent variable matrix}

\item{Z_mat}{external variable matrix}

\item{W}{cross-sectionally invariant variables - not used now}

\item{T}{time dimension of the data}

\item{N}{cross-sectional dimension of the data}

\item{aux_Y0}{auxiliary matrix with initial observations of the dependent variable}
}
\description{
Produce (negative) log marginal likelihood for QMLE with correlated random
coefficients
}
