% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{post_mean_lambda_par}
\alias{post_mean_lambda_par}
\title{Provide posterior means of lambda_i's based on the Parametric Posterior Mean
estimator with correlated random coefficients}
\usage{
post_mean_lambda_par(lambda0, sigma2, mmu, ww2_lambda, W, aux_Y0)
}
\arguments{
\item{lambda0}{initial estimate of lambdas}

\item{sigma2}{variance of the shocks}

\item{mmu}{auxiliary result (mean)}

\item{ww2_lambda}{auxiliary result (lambda times ww2)}

\item{W}{cross-sectionally invariant variables - not used now}

\item{aux_Y0}{auxiliary matrix with initial observations of the dependent variable}
}
\description{
Provide posterior means of lambda_i's based on the Parametric Posterior Mean
estimator with correlated random coefficients
}
