% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_summary.R
\name{summary.pmpp}
\alias{summary.pmpp}
\title{Summary method for objects of class \code{pmpp}.}
\usage{
\method{summary}{pmpp}(object, file = "", ...)
}
\arguments{
\item{object}{object of class \code{pmpp}, as returned by \code{pmpp()}}

\item{file}{a connection, or a character string naming the file to print to}

\item{...}{other parameters passed further}
}
\value{
A summary object for class \code{pmpp}.
}
\description{
Summary method for objects of class \code{pmpp}.
}
