% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{get_kernel}
\alias{get_kernel}
\title{Obtain 2D kernel density estimates given sufficient statistics for lambdas
and the initial data Y0}
\usage{
get_kernel(lambdas, sigma2, dens_grid, N, T, Y0)
}
\arguments{
\item{lambdas}{.}

\item{sigma2}{.}

\item{dens_grid}{.}

\item{N}{.}

\item{T}{.}

\item{Y0}{.}
}
\description{
Obtain 2D kernel density estimates given sufficient statistics for lambdas
and the initial data Y0
}
