% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{get_lambda0}
\alias{get_lambda0}
\title{Produce sufficient statistics (lambda0) given the common coefficients (rho0)}
\usage{
get_lambda0(rho, alpha = rep(0, n_alpha), N, T, n_alpha, Y_mat, X_mat, W,
  Z_mat)
}
\arguments{
\item{rho}{.}

\item{alpha}{.}

\item{N}{.}

\item{T}{.}

\item{n_alpha}{.}

\item{Y_mat}{.}

\item{X_mat}{.}

\item{W}{.}

\item{Z_mat}{.}
}
\description{
Produce sufficient statistics (lambda0) given the common coefficients (rho0)
}
