% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{get_sigma2}
\alias{get_sigma2}
\title{Produce variance of the shocks estimated using GMM residues (sigma2_0)
given the common coefficients (rho0)}
\usage{
get_sigma2(rho, alpha = 0, common_par_method, X_star, Y_star, Z_star, X_mat,
  Y_mat, Z_mat, n_alpha)
}
\arguments{
\item{rho}{.}

\item{alpha}{.}

\item{common_par_method}{.}

\item{X_star}{.}

\item{Y_star}{.}

\item{Z_star}{.}

\item{X_mat}{.}

\item{Y_mat}{.}

\item{Z_mat}{.}

\item{n_alpha}{.}
}
\description{
Produce variance of the shocks estimated using GMM residues (sigma2_0)
given the common coefficients (rho0)
}
