% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{GMM_parametric}
\alias{GMM_parametric}
\title{Produce posterior means of lambda's for the parametric GMM implementation
given autoregressive coefficient (rho)}
\usage{
GMM_parametric(rho, alpha = 0, optim_method, init, n_lambda, n_alpha, X_mat,
  Y_mat, Z_mat, W, T, N, aux_Y0, common_par_method, X_star, Y_star, Z_star)
}
\arguments{
\item{rho}{.}

\item{alpha}{.}

\item{optim_method}{.}

\item{init}{.}

\item{n_lambda}{.}

\item{n_alpha}{.}

\item{X_mat}{.}

\item{Y_mat}{.}

\item{Z_mat}{.}

\item{W}{.}

\item{T}{.}

\item{N}{.}

\item{aux_Y0}{.}

\item{common_par_method}{.}

\item{X_star}{.}

\item{Y_star}{.}

\item{Z_star}{.}
}
\description{
Produce posterior means of lambda's for the parametric GMM implementation
given autoregressive coefficient (rho)
}
