% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{loglikelihood_GMM}
\alias{loglikelihood_GMM}
\title{Produce negative log-likelihood in the GMM case}
\usage{
loglikelihood_GMM(theta, rho_GMMpar, alpha_GMMpar, sigma2_GMMpar, n_alpha,
  X_mat, Y_mat, Z_mat, W, T, N, aux_Y0)
}
\arguments{
\item{theta}{.}

\item{rho_GMMpar}{.}

\item{alpha_GMMpar}{.}

\item{sigma2_GMMpar}{.}

\item{n_alpha}{.}

\item{X_mat}{.}

\item{Y_mat}{.}

\item{Z_mat}{.}

\item{W}{.}

\item{T}{.}

\item{N}{.}

\item{aux_Y0}{.}
}
\description{
Produce negative log-likelihood in the GMM case
}
