% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{loglikelihood_QMLE}
\alias{loglikelihood_QMLE}
\title{Produce (negative) log marginal likelihood for QMLE with correlated random
coefficients}
\usage{
loglikelihood_QMLE(param, n_alpha, X_mat, Y_mat, Z_mat, W, T, N, aux_Y0)
}
\arguments{
\item{param}{.}

\item{n_alpha}{.}

\item{X_mat}{.}

\item{Y_mat}{.}

\item{Z_mat}{.}

\item{W}{.}

\item{T}{.}

\item{N}{.}

\item{aux_Y0}{.}
}
\description{
Produce (negative) log marginal likelihood for QMLE with correlated random
coefficients
}
