% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_plot.R
\name{plot.pmpp}
\alias{plot.pmpp}
\title{Plot method for objects of class \code{pmpp}.}
\usage{
\method{plot}{pmpp}(object, file = "", ...)
}
\arguments{
\item{object}{object of class \code{pmpp}, as returned by \code{pmpp()}}

\item{file}{a connection, or a character string naming the file to print to}

\item{...}{not used}
}
\value{
No object is returned. Displays a \code{ggplot} of density of the estimated 
individual-specific effects.
}
\description{
Plot method for objects of class \code{pmpp}.
}
\examples{
data(EmplUK, package = "plm")
EmplUK <- dplyr::filter(EmplUK, year \%in\% c(1978, 1979, 1980, 1981, 1982))
pmpp_model <- pmpp(dep_var = "emp", data = EmplUK)
plot(pmpp_model)
}
