% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmpp_aux_funcs.R
\name{post_mean_lambda_par}
\alias{post_mean_lambda_par}
\title{Provide posterior means of lambda_i's based on the Parametric Posterior Mean
estimator with correlated random coefficients}
\usage{
post_mean_lambda_par(lambda0, sigma2, mmu, ww2_lambda, W, aux_Y0)
}
\arguments{
\item{lambda0}{.}

\item{sigma2}{.}

\item{mmu}{.}

\item{ww2_lambda}{.}

\item{W}{.}

\item{aux_Y0}{.}
}
\description{
Provide posterior means of lambda_i's based on the Parametric Posterior Mean
estimator with correlated random coefficients
}
